python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
playwright==1.50.0
google-cloud-storage==2.14.0
jinja2==3.1.3